% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize.R
\name{design_optimization}
\alias{design_optimization}
\title{Evaluate the performance of alternative pooled testing designs}
\usage{
design_optimization(
  s.max = 32,
  w.max,
  p,
  n.max = w.max * s.max,
  w.min = 1,
  s.min = 1,
  max.groups = 50,
  score.fun = function(grid) {     with(grid, log(err/p) + log(unc/p) + log(base.unc/p)
    + log(abs(base.prev - p)/p) + log(w/s)) },
  a = 0.3,
  b = 0.3,
  ...
)
}
\arguments{
\item{s.max}{Maximum number of individual samples in a pool. To be chosen
after internal tests to evaluate possible loss in sensitivity due to
pooling.}

\item{w.max}{Maximum number of tests which are considered affordable for the
study.}

\item{p}{Maximal expected prevalence in the target population. It is better
to chose a value higher than what it is really expected in order not to
risk to design an underpowered study.}

\item{n.max}{Total number of individual samples that can be collected}

\item{w.min}{Minimum number of tests which is possible to perform.}

\item{s.min}{Minimum pool size for pooling.}

\item{max.groups}{The maximum number of rules that should be identified (the
actual number may be lower).}

\item{score.fun}{The function that computes the score (lower is better). It
takes as argument grid, wich is the simulated results plus enrichment via
\code{enrich_simulation()}.}

\item{a}{The \eqn{alpha} parameter of the prior distribution.}

\item{b}{The \eqn{beta} parameter of the prior distribution.}

\item{...}{Arguments to passed to \code{ctree()}, to control the creation of
the rules}
}
\value{

}
\description{
This tool conduct a grid search among possible pairs of pool sizes (number of
samples per pool) and total number of tested pools, given some operational
limits and an hypothetical underlying disease prevalence. Then for each pair
the algorithm simulate a result using ML and back-estimate the prevalence.
From the estimates the function computes a log score based on the estimation
error and the estimation uncertainty both for the pooled and unpooled (pool
size = 1) estimation. The score is also increased by a term given by the
ratio of the number of pools to the pool size. This last term act as balance
to precise estimations and lab resource usage. Finally a decision tree model
partitions the simulated solution in optimization windows defined by a range
of pool sizes and total number of tests. These windows can help the users to
choose the best study design for their specific setting.
}
\details{
The suggested optimal designs depend strongly on the expected disease
prevalence. In general a lower prevalence allows to have larger pool sizes
without loss of estimation accuracy and therefore larger saving of testing
resources. It is suggested to choose a prevalence higher than the expected
one (e.g. slightly higher than the expected higher possible value) in order
not to end up with an underpowered study in case the prevalence is higher
than expected,
}
\examples{

# Explore possible test design given a maximum pool size of 12, 2000 maximum
# number of tests and 2000 sampled individuals and a 5\% expected prevalence

grid <- design_optimization(s.max = 12, w.max = 2000, p = .05, n.max = 2000)

plot_optimization_grid(grid)
}
