% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{enrich_simulation}
\alias{enrich_simulation}
\title{Enrich simulated estimations with accuracy information}
\usage{
enrich_simulation(simulations, a = 0.3, b = 0.3)
}
\arguments{
\item{simulations}{A data frame with as columns the pool size \code{s}, the
number of pools \code{w}, the number of positive cases \code{cases}, the
prevalence estimate \code{Est}, the upper \code{Up} and lower \code{Lo}
uncertainty the expected real prevalence \code{p}.}

\item{a}{\eqn{\alpha} parameter of the prior Beta distribution for the
unpooled estimate.}

\item{b}{\eqn{\beta} parameter of the prior Beta distribution for the
unpooled estimate.}
}
\value{
A data frame with added the uncertainty size \code{unc} and the
estimation error \code{err}. Also the estimated prevalence in an unpooled
study \code{base.prev} is reported, together with the unpooled uncertainty
\code{base.unc} and error \code{base.err}.
}
\description{
Given some simulated estimates, compute the uncertainty interval size and
estimation error given a expected real prevalence. It is also computed the
expected prevalence in an unpooled study, plus the relative estimation error
and uncertainty size. The unpooled prevalence is computed using the Conjugate
Bayes method, with Beta prior parameters both equal to 0.3
}
\examples{
# Enrich estimations from simulated studies with accuracy information

set.seed(1234)
simulations <- simulate_pool_test(s = 8, w = 250, p = .01, iter = 3, consider.sensitivity = FALSE)
enrich_simulation(simulations)

}
