% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{evaluate_beta_params}
\alias{evaluate_beta_params}
\title{Inspect the shape of a beta distribution described by the \eqn{\alpha} and
\eqn{\beta} arguments}
\usage{
evaluate_beta_params(
  alpha,
  beta,
  lower.bound = 0.025,
  upper.bound = 0.975,
  plot = T,
  cut.plot.at = NA
)
}
\arguments{
\item{alpha}{\eqn{\alpha} Shape parameter of the Beta distribution.}

\item{beta}{\eqn{\beta} Shape parameter of the Beta distribution.}

\item{lower.bound}{Lower bound of the distribution to report.}

\item{upper.bound}{Upper bound of the distribution to report.}

\item{plot}{Whether to plot the PDF of the distribution.}

\item{cut.plot.at}{Since distribution defined by \code{alpha} or \code{beta}
< 1 are strongly concentrated at 0 or 1, squeezing down the rest of the PDF
in the plot, it may be useful to cut the y axis in order to show the
interesting parts of the distribution.}
}
\value{
A vector with the lower/upper bound chosen by the user, the median,
the mean, the standard deviation and the mode of the distribution. If
\code{alpha} or \code{beta} are below 1, the mode cannot be calculated
through \code{(alpha - 1)/(alpha + beta - 2)} so it is approximated through
grid search
}
\description{
This function is useful in define priors based on the beta distribution.
Given the \eqn{\alpha} and \eqn{\beta} parameters, it returns relevant
statistics of the distribution and boundary points to help understand its
characteristics. Optionally the full PDF of the distribution is plotted.
}
\examples{

# \code{Alpha} and \code{beta} chosen in order to have the mean
# \code{alpha/(alpha + beta)} equal to .05 and precision \code{alpha + beta}
# equal to 30

evaluate_beta_params(.05 * 30, (1 - .05) * 30)

# This time the \code{alpha} and \code{beta} are chosen to have a mode
# \code{(alpha - 1)/(alpha + beta - 2)} equal to 0.5, with the same precision

evaluate_beta_params(.05 * (30 - 2) + 1, (1 - .05) * (30 - 2) + 1)

# If we again center the mean at 0.05 but this time the precision is lowered to 15,
# \code{alpha} goes below 1 and this produces a distribution with most of the mass
# at zero. The mode value calculated analitycally in this case would be wrong (negative)
# so the empirical mode is shown as an approximation

evaluate_beta_params(.05 * 15, (1 - .05) * 15)



}
