% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{format_simulation}
\alias{format_simulation}
\title{Formats the results of a simulation}
\usage{
format_simulation(simulations, level = 0.95, reverse = T)
}
\arguments{
\item{simulations}{The results of a call to \code{simulate_pool_test}}

\item{level}{The uncertainty level around the estimates.}

\item{reverse}{Whether to show the results in vertical format instead of wide.}
}
\value{
A data frame with reformatted titles
}
\description{
Adds more readable names, transforms estimates to percentages and adds uncertainty
intervals around the simulated parameters. If the simulation are not already enriched
via \code{enrich_simulation} it does it automatically.
}
\examples{

sim <- simulate_pool_test(s = 8, w = 250, p = .01, iter = 3000, consider.sensitivity = FALSE)

formatted.sim <- format_simulation(sim)

}
