% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_beta_params}
\alias{get_beta_params}
\title{Estimate Beta distibution parameters given a mean and uncertainty boundaries}
\usage{
get_beta_params(
  min.val,
  mid.val,
  max.val,
  levels = c(0.025, 0.975),
  mid.val.type = c("mean", "median", "mode"),
  force.proper = TRUE,
  plot = FALSE
)
}
\arguments{
\item{min.val}{Expected lower uncertainty bound of the ditribution.}

\item{mid.val}{Expected central value. Can represent the median or the mean
of the distribution according to the \code{mid.val.type} parameter.}

\item{max.val}{Expected upper uncertainty bound of the ditribution.}

\item{levels}{Uncertainty level corresponding to the \code{min.val} and
\code{max.val} parameters.}

\item{mid.val.type}{Define which parameter of the distribution \code{mid.val}
refers to, either the mean, the median or the mode.}

\item{force.proper}{Discourage values of alpha and beta < 1 during the
optimization process. Tests should be done turning this option on and off
and evaluating the results}

\item{plot}{Whether to plot the resulting distribution using
\code{\link{evaluate_beta_params}} fuction.}
}
\value{
A list with the proposed alpha and beta parameters, a description of
the resulting distribution through \code{\link{evaluate_beta_params}}, and
the actual output of the optimization algorithm.
}
\description{
This tool may be helpful in building priors based on the Beta distribution.
The user need to pass the expected mean/median of the distibution and the
uncertainty boundaries. Then an optimizion algorithm is used to the find the
alpha and beta parameters of the beta that reproduce a shape as similar as
possible to the expected one. The method don't assure perfect coverage,
either for optimization limits or because the expected shape is not allowed
by the Beta distribution, but is a starting point for further esploration of
useful alfa and beta parameters.
}
\examples{

# Estimate the alpha and beta parameter of a Beta distribution with 95\%
# boundaries near 0.0001 and 0.1 and mean value 0.015

get_beta_params(.0001, .015, .1, force.proper = TRUE, plot = TRUE)

# Same inputs but now force.proper is FALSE

get_beta_params(.0001, .015, .1, force.proper = FALSE, plot = TRUE)

}
\seealso{
\code{\link{evaluate_beta_params}}
}
