% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulate_pool_test}
\alias{simulate_pool_test}
\title{Simulation of pooled prevalence studies}
\usage{
simulate_pool_test(
  s = 12,
  w = 200,
  p = 0.01,
  iters = 3000,
  consider.sensitivity = FALSE,
  simulate.results = TRUE,
  asens = 8.88,
  bsens = 0.74,
  estimation.method = "B"
)
}
\arguments{
\item{s}{Number of sampled individuals per pool (default \code{12})}

\item{w}{Number of pools to test (default \code{200})}

\item{p}{Expected population prevalence (default \code{0.01})}

\item{iters}{Number of simulations (default \code{3000})}

\item{consider.sensitivity}{Consider sensitivity in sample acquisition in the
simulation (default \code{FALSE})}

\item{simulate.results}{Add estimation results to the simulation (default
\code{TRUE})}

\item{asens}{\eqn{\alpha} Parameter of the Beta distribution for the sampling
sensitivity (default \code{8.88})}

\item{bsens}{\eqn{\beta} Parameter of the Beta distribution for the sampling
sensitivity (default \code{0.74})}

\item{estimation.method}{Whether to use Bayesian Conjugate \\(default\\) or
MLE for the prevalence estimation (default \code{'CB'})}
}
\value{
A dataframe with: the initial w, s, and p parameters; p.sample: the
individual risk ditribution quantiles (0.025, .5, .975); cases: the number
of positive individuals in the sample, after accounting for sampling
sensitivity; pos & neg: the number of positive and negative pools out of w
tested pools.
}
\description{
Generate a sample of individuals with a given disease status informed by a
chosen underlying diseases prevalence and risk variability. Then simulate a
pooled test results from such sample. The process can be repeated for a
number of iteration to have a distribution of results. The prevalence in each
simulation is estimated through the \code{get_estimates(method = 'CB')}
function.
}
\examples{
# Simulate a study in a population with a disease risk of 1\%

set.seed(1234)
simulate_pool_test(s = 8, w = 250, p = .01, iter = 1, consider.sensitivity = FALSE)

# Same simulation this time considering false negatives due to sampling sensitivity,
# with \code{asens} and \code{bsens} chosen to have a mean sensitivity of 95\% and
# a lower bound around 70\%

set.seed(1234)
simulate_pool_test(s = 8, w = 250, p = .01, iter = 1, consider.sensitivity = TRUE)

# use the \code{iters} argument to create more simulations

simulate_pool_test(s = 8, w = 250, p = .01, iter = 3000, consider.sensitivity = FALSE)

}
