% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{get_estimates}
\alias{get_estimates}
\title{Prevalence estimation from a pooled study}
\usage{
get_estimates(
  s,
  w,
  k = NULL,
  p_test = NULL,
  p = NULL,
  level = 0.95,
  method = c("BC", "HC", "ML"),
  a = 0.3,
  b = 0.3,
  consider.sensitivity = T,
  asens = 8.88,
  bsens = 0.74
)
}
\arguments{
\item{s}{A vector of pool sizs (number of individual samples in a pool)}

\item{w}{A vector of numbers of tested pools in a study}

\item{k}{A vector of numbers of positive pools in a study}

\item{p_test}{A vector of probabilities of positive pools in a study. if
\code{k} is not provided, it used to generate the expected number of
positive pools}

\item{p}{A vector of prevalences of the disease. if \code{k} and
\code{p_test} is not provided, it used to generate the expected number of
positive pools}

\item{level}{The uncertainty level to use for the confidence/credibility
interval. Must be strictly greater than 0 and less than 1. Defaults to
0.95, which corresponds to a 95 percent confidence interval}

\item{method}{Either Maximum Likelihood Estimation (ML), Conjugate Bayesian
(CB) estimation, Hierarchical MCMC Bayesian estimation (HB)}

\item{a}{\eqn{\alpha} parameter of the Beta prior. The prior is on the test
results in case of \code{method = 'CB'} and on the prevalence estimation in
case of \code{method = 'HB'}}

\item{b}{\eqn{\beta} parameter of the Beta prior. The prior is on the test
results in case of \code{method = 'CB'} and on the prevalence estimation in
case of \code{method = 'HB'}}

\item{asens}{\eqn{\alpha} parameter of the Beta distribution describing the
sample acquisition sensitivity}

\item{bsens}{\eqn{\beta} parameter of the Beta distribution describing the
sample acquisition sensitivity}
}
\value{
A named numeric matrix of height equal to the length of the parameter
vectors. The matrix contains: \describe{ \item{p_est}{The positivity rate
for the pools in the study} \item{k}{The number of positive pools}
\item{Est}{The estimated disease prevalence} \item{Lo}{The lower
uncertainty bound} \item{Up}{The upper uncertainty bound} }
}
\description{
The function allow to estimate prevalence from the results of prevalence
study. It use three statistical frameworks: Maximum Likelihood estimation
(ML), Conjugate Bayesian (CB) estimation, Hierarchical MCMC Bayesian
estimation (HB).
}
\details{
The CB and HB estimation allows the setting of priors, the first on the study
results, the second on the prevalence estimates themselves; the default prior
is a Beta(0.5, 0.5) distribution to account for the positive bias in the
prevalence estimation in case of low sample size. The HB method allows also
to take in account the sensitivity of the sample acquisition procedure,
distibuted as ~ Beta(asens, bsens). The default parameters describe a mean
sensitivity of 95\% with a lower bound near 90\%.

The function takes in input the pool size, the number of tested pools and the
positive pools. In alternative, to the actual test results, users can provide
only the probability of positive pools in the study or the theoretical mean
prevalence; the expected number of positive pools is then generated throgh
Maximum Likelihood. All parameters study parameters \code{s, w, p_tests, p}
can be vectors; this permit to analyze multiple (real or simulated) studies
at once.

The HB method requires the compilation of a STAN model at the first use. The
model will be compiled only the first time.
}
\examples{

# Estimate prevalence from a study with 30 positive pools out of 200, with 10 samples per pool

get_estimates(s = 10, w = 200, k = 30)

# Same study now analyzed with a full bayesian estimation considering sampling sensitivity.
# Notice the slightly higher estimated prevalence, due to the incorporation of the false negatives due to sampling

\dontrun{
   get_estimates(s = 10, w = 200, k = 30, method = 'HC')
}


}
